Rename feature.
