+++ Cargo.toml
@@ -14,7 +14,7 @@ license = "MIT"
 
 [features]
 32bit = []
-tokio = ["dep:tokio", "dep:async-trait"]
+async = ["dep:tokio", "dep:async-trait"]
 
 [dependencies]
 tokio = { version = "1", optional = true, default-features = false, features = ["io-util"] }
